fix(preview): keep iframe state across mode toggles + DevTools-grade picker
Issue #2143 — toggling Comment / Inspect / Draw / Edit / Tweaks remounted
the artifact iframe and reset multi-page React / multi-HTML prototypes
back to the entry file. Selection was also broken on un-annotated
artifacts and on micro-components inside annotated cards.

Changes:

- packages/contracts/src/preview-bridges: new shared module exporting
  SELECTION_BRIDGE_SCRIPT, SELECTION_BRIDGE_STYLE, ROUTE_PERSIST_SCRIPT
  as plain HTML fragment strings. Imported by both daemon and web so
  the URL-load and srcDoc transports run identical bridges.

- apps/daemon/src/project-routes.ts: when serving HTML from
  /api/projects/:id/raw/*, splice in the three bridges. Idempotent via
  marker attributes; the script-tag check uses the exact opener
  '<script data-od-selection-bridge>' to avoid the style block's
  '-style' suffix false-positiving the substring match.

- apps/web/src/runtime/srcdoc.ts: replace ~600 lines of inline
  selection-bridge body with imports from the shared module. Add
  injectRoutePersistBridge() so srcDoc-loaded previews also persist
  SPA routes across mode-toggle remounts.

- apps/web/src/components/file-viewer-render-mode.ts:
  - hasUrlModeBridge() now also matches '<script data-od-selection-bridge>'
    so the daemon-injected bridge counts as urlModeBridge:true.
  - Inspect mode no longer hard-forces srcDoc when the URL-load bridge
    is present (selection bridge handles inspect picking too; only the
    inspect override sheet is bridge-internal).
  - Edit mode always forces srcDoc — the URL-load bridge ships
    selection only, not the manual-edit bridge.

- packages/contracts/{package,esbuild.config}.{json,mjs}: wire the new
  './preview-bridges' subpath export so workspace consumers can import
  it as @open-design/contracts/preview-bridges.

- apps/web/src/components/FileViewer.tsx: rewrite the inspect/comment
  empty-hint copy. Picker mode now reads "Hover any element to
  highlight it, then click to leave a comment." — accurate for the
  DevTools-style behavior, no longer asks for data-od-id annotations
  the user can't add.

Bridge behavior changes:

- Picker mode is leaf-first. closestTarget() walks UP from event.target
  and the deepest visible element wins, regardless of any annotated
  ancestor. Mirrors how Chrome/Firefox inspect picks the element under
  the cursor. Inspect mode keeps the upstream annotated-ancestor-wins
  behavior so per-selector CSS overrides have stable selectors.

- meaningfulDomFallbackTarget() reduced to "is visible" — DevTools
  matches every visible element under the cursor; we do too. The
  closestTarget walk handles the structural decisions.

- canUseDomFallback() drops the "zero annotations" gate. Mixed
  artifacts (annotated main pane + un-annotated nav sidebar) used to
  leave the nav un-pickable; now the inner walk prefers annotated
  ancestors when they exist and falls back to leaf otherwise.

- Route-persist honors the host's '?r=N' reload generation. When the
  user clicks the toolbar Reload button the URL gets a new 'r' value;
  on the next mount the bridge sees the bumped generation and clears
  saved route so reload actually returns to the entry file. Mode-
  toggle remounts reuse the same '?r=' value and restore the saved
  route as before.

Verified end-to-end against a live container with Playwright /
Chromium (see /home/sarsingm/SKILLER/open-design-issue-2143-fix.md
gotchas + verification recipes). All six checks pass:

  * SELECTION_BRIDGE_SCRIPT, SELECTION_BRIDGE_STYLE, ROUTE_PERSIST_SCRIPT
    all present in served HTML.
  * window.__odSelectionBridgeInstalled === true,
    window.__odRoutePersistInstalled === true after iframe load.
  * od:comment-mode enabled=true → htmlAttr toggles, od:comment-targets
    emits non-empty list (110 targets on hr-xpress/screens/help.html).
  * Hovering the big-number leaf (.kpi-value / span.num) emits
    od:comment-hover with the LEAF as label, not the annotated card
    ancestor.
  * Click on a nav <a> in comment mode emits od:comment-target with
    freePin=false, iframe URL does not change.
  * Force a same-?r remount → iframe stays on
    /screens/help.html?x=42. Bump ?r=N+1 → returns to entry file.

Doc with annotated patch + Playwright recipes:
  /home/sarsingm/SKILLER/open-design-issue-2143-fix.md
