+++ apps/web/src/components/file-viewer-render-mode.ts
@@ -69,10 +69,16 @@ export function shouldUrlLoadHtmlPreview(d: UrlLoadDecision): boolean {
   if (d.mode !== 'preview') return false;
   if (d.isDeck) return false;
   if (d.commentMode && !d.urlModeBridge) return false;
-  // Inspect needs the selection bridge injected via buildSrcdoc; a raw
-  // URL-loaded iframe has no listener to apply per-element overrides.
-  if (d.inspectMode) return false;
-  if (d.editMode && !d.urlModeBridge) return false;
+  // Issue #2143 — when the URL-loaded HTML carries our injected selection
+  // bridge (daemon-side, see project-routes.ts injectPreviewBridgesIntoHtml),
+  // inspect mode can stay on URL-load too. Without the bridge the host has
+  // no listener to apply per-element overrides, so we still fall back to
+  // srcDoc the way upstream does.
+  if (d.inspectMode && !d.urlModeBridge) return false;
+  // Manual edit needs the edit-mode bridge which is srcDoc-only — the
+  // daemon-injected URL-load bridge ships selection only, not manual edit.
+  // Always fall back to srcDoc when editMode is on.
+  if (d.editMode) return false;
   // Palette tweaks need the srcDoc-side bridge — `<iframe src=URL>` has
   // no parent-injected listener to recolor against.
   if (d.paletteActive) return false;
@@ -86,9 +92,19 @@ export function shouldUrlLoadHtmlPreview(d: UrlLoadDecision): boolean {
   return true;
 }
 
+// Recognizes either the artifact-owned `od-direct-edit.js` (legacy) or the
+// daemon-injected `<script data-od-selection-bridge>` (Issue #2143). Both
+// give the host a postMessage target for picker/inspect inside URL-load
+// previews, so the render-mode decision can keep URL-load on for comment
+// and inspect mode without flipping to srcDoc.
+//
+// The selection-bridge marker requires the immediate-`>` form so the style
+// tag's `data-od-selection-bridge-style` suffix doesn't false-positive.
 export function hasUrlModeBridge(source: string | null | undefined): boolean {
   if (!source) return false;
-  return /<script\b[^>]*\bsrc\s*=\s*["'][^"']*\bod-direct-edit\.js\b[^"']*["'][^>]*>/i.test(source);
+  if (/<script\b[^>]*\bsrc\s*=\s*["'][^"']*\bod-direct-edit\.js\b[^"']*["'][^>]*>/i.test(source)) return true;
+  if (/<script\s+data-od-selection-bridge\s*>/i.test(source)) return true;
+  return false;
 }
 
 /**
